+++ src/com/app/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package com.app.calculator.view;
+package com.app.calculator;
 
 import java.util.List;
 
@@ -7,6 +7,9 @@ import javax.swing.JFrame;
 import javax.swing.JTextField;
 
 import com.app.calculator.model.CalculatorInputVO;
+import com.app.calculator.view.Frame;
+import com.app.calculator.view.TextField;
+import com.app.calculator.view.button.Buttons;
 
 /**
  *
